+++ apps/AlbionMC_API/src/core/entities/crafting_slot.py
